+++ src/client/types.ts
@@ -1,3 +1,4 @@
+import { MessageType } from '@src/commands/types'
 import { ChatUserstate } from 'tmi.js'
 import { ChatPermissionLevel, GlobalPermissionLevel } from '../utilities/permission/types'
 
@@ -7,4 +8,18 @@ export interface TwitchUserState extends ChatUserstate {
 export interface BotResponse {
   success?: boolean
   response: string | string[]
+}
+
+export interface ChatContext {
+  where: string
+  type: MessageType
+  self: boolean
+  message: string
+  user: TwitchUserState
+}
+
+export interface ResponseContext {
+  where: string
+  response: BotResponse
+  type: MessageType
 }