+++ CHANGELOG.md
@@ -13,12 +13,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - hot reload for dev
 - emote command to fetch all emotes for channel
 - save emotegame results as stats
+- commands can now be run via whispers
+
 ### Changed
 
 - remove channel from Response
 - use database-url instead of single params
 - ping command returns more information
 - response can be string array or string now
+
 ### Fix
 
 - no longer sending letters after word has been guessed